# Optimize System.HexConverter.IsHexChar on 64 bits